improve package structure and fiy some bugs
